- Fixed some wxWidgets warnings - Added splash window+progress bar when opening the texture editor (can take a VERY long time to open when there are a lot of patches :/ Need to speed up entry searching)
git-svn-id: http://mancubus.net/svn/hosted/slade/trunk@752 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/ArchiveManagerPanel.cpp
@@ -602,7 +602,7 @@ void ArchiveManagerPanel::openFile(string filename) {
 	Archive* new_archive = theArchiveManager->openArchive(filename);
 
 	sw.Pause();
-	wxLogMessage("Opening took %d ms", sw.Time());
+	wxLogMessage("Opening took %d ms", (int)sw.Time());
 
 	// Hide splash screen
 	theSplashWindow->hide();
